# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_yourusername_noshirvani-academy_backend_internal_domain.Exam:
    properties:
      created_at:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      dynamic_fields:
        additionalProperties: true
        type: object
      exam_date:
        type: string
      id:
        type: string
      jalali_date:
        type: string
      major:
        type: string
      student:
        $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Student'
      student_id:
        type: string
      subjects:
        items:
          $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.SubjectExam'
        type: array
      title:
        type: string
      total_subjects:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_yourusername_noshirvani-academy_backend_internal_domain.Mistake:
    properties:
      category:
        type: string
      created_at:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      dynamic_fields:
        additionalProperties: true
        type: object
      exam:
        $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Exam'
      exam_id:
        type: string
      id:
        type: string
      notes:
        type: string
      question_number:
        type: integer
      student:
        $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Student'
      student_id:
        type: string
      subject_exam:
        $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.SubjectExam'
      subject_exam_id:
        type: string
      updated_at:
        type: string
    type: object
  github_com_yourusername_noshirvani-academy_backend_internal_domain.Student:
    properties:
      approval_date:
        type: string
      birth_date:
        type: string
      city:
        type: string
      created_at:
        type: string
      dynamic_fields:
        additionalProperties: true
        type: object
      first_name:
        type: string
      id:
        type: string
      is_approved:
        type: boolean
      jalali_birth_date:
        type: string
      last_name:
        type: string
      major:
        type: string
      profile_photo:
        type: string
      school:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.User'
      user_id:
        type: string
    type: object
  github_com_yourusername_noshirvani-academy_backend_internal_domain.SubjectExam:
    properties:
      answered:
        type: integer
      blank:
        type: integer
      correct:
        type: integer
      exam_id:
        type: string
      id:
        type: string
      percentage:
        type: number
      subject_name:
        type: string
      total_questions:
        type: integer
      wrong:
        type: integer
    type: object
  github_com_yourusername_noshirvani-academy_backend_internal_domain.User:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      phone:
        type: string
      role:
        type: string
      student:
        $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Student'
      updated_at:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handler.CreateExamInput:
    properties:
      dynamic_fields:
        additionalProperties: true
        type: object
      exam_date:
        type: string
      jalali_date:
        example: 1400/01/01
        type: string
      major:
        type: string
      subjects:
        items:
          $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.SubjectExam'
        type: array
      title:
        type: string
      total_subjects:
        type: integer
    type: object
  internal_handler.CreateMistakeInput:
    properties:
      category:
        type: string
      dynamic_fields:
        additionalProperties: true
        type: object
      exam_id:
        type: string
      notes:
        type: string
      question_number:
        type: integer
      subject_exam_id:
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@noshirvaniacademy.com
    name: تیم توسعه اپلیکیشن
    url: https://github.com/Offline-Devs
  description: |-
    API برای سیستم مدیریت آکادمی نوشیروانی
    شامل خدماتی برای احراز هویت، مدیریت دانشجویان، آزمون‌ها، اشتباهات و مقالات بلاگ
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: نوشیروانی آکادمی API
  version: "1.0"
paths:
...
  /exams:
    get:
      description: تمام آزمون‌های دانشجو را دریافت می‌کند
      produces:
      - application/json
      responses:
        "200":
          description: لیست آزمون‌ها
          schema:
            items:
              $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Exam'
            type: array
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: پروفایل دانشجو یافت نشد
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: دریافت لیست آزمون‌های دانشجو
      tags:
      - آزمون‌ها
    post:
      consumes:
      - application/json
      description: یک آزمون جدید برای دانشجو ایجاد می‌کند
      parameters:
      - description: اطلاعات آزمون
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateExamInput'
      produces:
      - application/json
      responses:
        "201":
          description: آزمون با موفقیت ایجاد شد
          schema:
            $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Exam'
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: پروفایل دانشجو یافت نشد
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ایجاد آزمون جدید
      tags:
      - آزمون‌ها
  /exams/{id}:
    delete:
      description: یک آزمون را حذف می‌کند
      parameters:
      - description: شناسه آزمون
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: آزمون با موفقیت حذف شد
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: پروفایل دانشجو یافت نشد
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: حذف آزمون
      tags:
      - آزمون‌ها
    get:
      description: جزئیات یک آزمون خاص را دریافت می‌کند
      parameters:
      - description: شناسه آزمون
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: جزئیات آزمون
          schema:
            $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Exam'
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: آزمون یافت نشد
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: دریافت جزئیات آزمون
      tags:
      - آزمون‌ها
...
  /mistakes:
    get:
      description: تمام اشتباهات ثبت‌شده برای دانشجو را دریافت می‌کند
      produces:
      - application/json
      responses:
        "200":
          description: لیست اشتباهات
          schema:
            items:
              $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Mistake'
            type: array
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: پروفایل دانشجو یافت نشد
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: دریافت لیست اشتباهات دانشجو
      tags:
      - اشتباهات
    post:
      consumes:
      - application/json
      description: یک اشتباه جدید برای دانشجو ثبت می‌کند
      parameters:
      - description: اطلاعات اشتباه
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateMistakeInput'
      produces:
      - application/json
      responses:
        "201":
          description: اشتباه با موفقیت ایجاد شد
          schema:
            $ref: '#/definitions/github_com_yourusername_noshirvani-academy_backend_internal_domain.Mistake'
        "400":
          description: درخواست نامعتبر
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: پروفایل دانشجو یافت نشد
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ایجاد اشتباه جدید
      tags:
      - اشتباهات
  /mistakes/{id}:
    delete:
      description: یک اشتباه را حذف می‌کند
      parameters:
      - description: شناسه اشتباه
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: اشتباه با موفقیت حذف شد
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: عدم اجازه دسترسی
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: پروفایل دانشجو یافت نشد
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: خطای سرور
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: حذف اشتباه
      tags:
      - اشتباهات
...
securityDefinitions:
  BearerAuth:
    description: '"توکن JWT - فرمت: Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"